requirements.txt
python-telegram-bot[job-queue]
ccxt==4.2.82
setuptools
pandas
requests